agentscope_project/requirements.txt
agentscope==1.0.0
flask==3.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
pillow==10.1.0
httpx==0.25.1
pywencai==0.13.1
pymysql==1.1.0